% Generated by roxygen2 (4.0.1): do not edit by hand
\name{workbench}
\alias{workbench}
\title{Explora germplasm selection tool}
\usage{
workbench()
}
\description{
Bioversity/CCAFS Seeds4Needs Explora germplasm selection tool
}
\details{
The main purpose of Explora is to select a set of promising accessions
from genebank collections that meet the interests of the user. Explora allows
1) selection of germplasm from big characterization and/or evaluation data sets;
2) consideration of more than one trait and trade-offs between different traits of interests; and
3) maximization of diversity for specific traits of interests.
}
\examples{
workbench()
}
\author{
Maarten van Zonneveld,  Johannes Ospina and Richard Bruskiewich
}

